Added support for DVB_drag plugin and some mappings
Added support for DVB_drag plugin
Added mapping for Q to avoid Ex mode and make it easier to get
   into viual block mode at the same time.
Added v mapping for J (I never want to join lines when in visual mode).
Added quick comments for dosbatch files.
Added Color colum autocmd to hilight the 81st character on the line if it
   exists.