Updated documentation plus some small additions
Added <A-a> mapping to do a vsplit to the right. (It always annoyed me
  that splits opened on the left, but I want my difsplits to open on the
  left. I'm just weird like that.)
Updated <F6> to be less generic but handle documenting vim functions very
  nicely. Creates a whole template for the function now.
When resourcing my vimrc it will now resource .vimpref too.
Updated SaveBuffer to use `quit` instead of feedkeys.
Updated ParagraphToEightyChars to only Join lines that are already
  comments.
Added the g:Tumbler_vimrc variable so I can tell if I've loaded it before