Various improvements
Added NMakefile to wildignore.
Uncommented Focus autocmds. (Had to add something to take care of
   Helplink.vim later in a different part.)
Added <A-BS> to command line
Made my custom K command only work on windows because 'man' is WAY
   too useful.
Added visual mapping for <A-j> and <A-k>
Added an autocmd to tell helplink.vim which registers to copy to
   allowing me to keep my custom Focus autocmds.
Moved setting line endings to a function so that it can be more
   easily overridden.
Added Python comments.
Updated paragraphToEightChars to handle some more edge cases.
Made some changes to the tips and tricks section.